refactor(backend): split oversized assistant session and extension services (#822)
